Add `yargs` module.
This replaces the unmaintained `optimist` module.  For continuity, we're
adding `yargs` without removing `optimist` yet.  This will let us merge
the Parsoid patch which replaces optimist with yargs, and then we can
make a new commit in the deploy repo which advances the `src` submodule
to the version which doesn't require `optimist` and removes `optimist`
from node_modules at the same time.

Change-Id: Iac1335af88336750ce2e594467158b999a182f9e