Conf: Add the production enwiki to the list of available domains
With the deprecation of the Parsoid's public entry point, RESTBase now
uses the labs instance to conduct proper tests. However, because the
tests are tied to a domain, and different series of tests exercise
different properties, it is a requirement for us to be able to retrieve
content for production's enwiki from this parsoid instance.

Note: as this is to be used by RESTBase internally only, we should not
advertise this.

Change-Id: I884a23a7e71b0c67a0ddbdd8b75f9cc1cea5fb53