Logger conf: output to a file instead of stdout
After the move to service-runner and service::node, regular users are
not able to see upstart logs. So, instead of making exceptions in the
shared services' code for Parsoid, have it output the logs into a file.
This is a temporary measure anyways, as with SystemD we are able to
redirect std streams to files.

This change has already been applied manually to beta and has proved to
work.

Bug: T90668
Change-Id: I71ab291e333d6d5519852648d4e56e30f94e43cd