Add the Scap3 configuration
This patch adds the Scap3 configuration necessary to be able to deploy
Parsoid using Scap3. It configures the deployment as a two-step process:
first, two nodes from each DC are deployed to (the canary nodes), and
the second phase proceeds to deploy to the remaining node (provided the
canary deployment succeeded).

Bug: T120103
Change-Id: Ibb82efe85c491850a4b0e8e2ead3bf4cd3d92077