Depool and repool nodes and set the group size to 7
During a deploy, we don't want the target nodes to be accepting new
connections, as these will get reset. So, after the new code has been
deployed, depool the servers, restart the service on them and finally
repool them.

Also, execute the deploy in batches of 7, so that the majority of nodes
is still pooled and can respond to requests.

Bug: T120103
Change-Id: Id2f49733c748915d19764d66cd8f552dec109913